requirements.txt
astral==1.6
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5.1
decorator==4.2.1
Django==2.0.3
django-configurations==2.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-extensions==2.0.3
django-leaflet==0.23.0
django-model-utils==3.1.1
djangorestframework==3.7.7
et-xmlfile==1.0.1
future==0.16.0
geocoder==1.38.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.11.1
mock==2.0.0
more-itertools==4.2.0
mpmath==1.0.0
numpy==1.14.2
openpyxl==2.5.1
parso==0.1.1
pbr==4.1.0
peewee==3.1.5
pexpect==4.4.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
py==1.5.4
Pygments==2.2.0
pytest==3.6.2
pytest-cov==2.5.1
pytest-django==3.3.2
pytest-djangoapp==0.7.1
pytz==2018.3
ratelim==0.1.6
requests==2.18.4
requests-cache==0.4.13
scipy==1.0.0
selenium==3.13.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
sympy==1.1.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
